+++ mitra_models/src/users/queries.rs
@@ -551,6 +551,25 @@ pub async fn create_portable_user(
     Ok(user)
 }
 
+pub async fn get_portable_user_by_name(
+    db_client: &impl DatabaseClient,
+    username: &str,
+) -> Result<PortableUser, DatabaseError> {
+    let maybe_row = db_client.query_opt(
+        "
+        SELECT portable_user_account, actor_profile
+        FROM portable_user_account JOIN actor_profile USING (id)
+        WHERE actor_profile.username = $1
+        ",
+        &[&username],
+    ).await?;
+    let row = maybe_row.ok_or(DatabaseError::NotFound("user"))?;
+    let db_user: DbPortableUser = row.try_get("portable_user_account")?;
+    let db_profile: DbActorProfile = row.try_get("actor_profile")?;
+    let user = PortableUser::new(db_user, db_profile)?;
+    Ok(user)
+}
+
 #[cfg(test)]
 mod tests {
     use serde_json::json;