+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implemented C2S inbox for portable users.
 - Put outgoing activities into inbox if portable recipient has local account.
 - Implemented C2S outbox for portable users.
+- Implemented Webfinger for portable users.
 
 ### Changed